+++ src/stream/src/executor/project_set.rs
@@ -187,68 +187,3 @@ impl ProjectSetExecutor {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use risingwave_common::catalog::{Field, Schema};
-    use risingwave_common::types::DataType;
-    use risingwave_expr::expr::build_from_pretty;
-    use risingwave_expr::table_function::repeat;
-
-    use super::super::test_utils::MockSource;
-    use super::super::*;
-    use super::*;
-    use crate::executor::test_utils::snapshot::executor_snapshot;
-    use crate::executor::test_utils::MessageSender;
-
-    const CHUNK_SIZE: usize = 1024;
-
-    fn create_executor() -> (MessageSender, BoxedMessageStream) {
-        let schema = Schema {
-            fields: vec![
-                Field::unnamed(DataType::Int64),
-                Field::unnamed(DataType::Int64),
-            ],
-        };
-        let (tx, source) = MockSource::channel(schema, PkIndices::new());
-
-        let test_expr = build_from_pretty("(add:int8 $0:int8 $1:int8)");
-        let tf1 = repeat(build_from_pretty("1:int4"), 1);
-        let tf2 = repeat(build_from_pretty("2:int4"), 2);
-
-        let project_set = Box::new(ProjectSetExecutor::new(
-            Box::new(source),
-            vec![],
-            vec![test_expr.into(), tf1.into(), tf2.into()],
-            1,
-            CHUNK_SIZE,
-        ));
-        (tx, project_set.execute())
-    }
-
-    #[tokio::test]
-    async fn test_project_set() {
-        let snapshot = executor_snapshot(
-            || std::future::ready(create_executor()),
-            r###"
-- !chunk |2
-      I I
-    + 1 4
-    + 2 5
-    + 3 6
-- !chunk |2
-      I I
-    + 7 8
-    - 3 6
-"###,
-        )
-        .await;
-
-        insta::with_settings!({
-            description => "($0+$1), repeat(1,1), repeat(2,2)",
-            omit_expression => true
-        }, {
-            insta::assert_snapshot!(snapshot)
-        });
-    }
-}
